Gets the translationHelps for the GL
